improvements to basic routes
