+++ src/API/REST/routes/get_profile.ts
@@ -4,19 +4,16 @@ import {userModel} from "../../../database/models/user";
 import check_auth from "../middleware/check_auth";
 
 export default function get_profile_route() {
-  let router = express.Router();
+    let router = express.Router();
 
-  router
-    .get("/profile/:id", check_auth(), async (req: ExtRequest, res: any) => {
-      const userId = req.params.id;
+    router.get("/profile/:id", check_auth(), async (req: ExtRequest, res: any) => {
+        const userId = req.params.id;
+        const user = await userModel.findById(userId);
 
-      const user = await userModel.findById(userId);
+        if (!user)
+            return res.status(404).send(`User ${userId} not found`);
 
-      if (!user) {
-        res.status(404).send(`User ${userId} not found`);
-      }
-
-      return res.status(200).json(user);
+        return res.status(200).json(user);
     });
-  return router;
+    return router;
 }
