+++ src/API/REST/routes/get_unauthorized_users.ts
@@ -4,20 +4,18 @@ import { ExtRequest } from "../../../definitions/ext_request";
 import check_auth from "../middleware/check_auth";
 
 export default function get_unauthorized_users_route() {
-  let router = express.Router();
+    let router = express.Router();
 
-  router
-    .get("/unauthorized_users", check_auth(), async (req: ExtRequest, res) => {
-      if (!req.is_administrator) {
-        return res.status(403).send("Unauthorized");
-      }
+    router.get("/unauthorized_users", check_auth(), async (req: ExtRequest, res) => {
+        if (!req.is_administrator)
+            return res.status(403).send("Unauthorized");
 
-      try {
-        const unauthorizedUsers = await userModel.find({isAuthorized: false});
-        return res.status(200).json(unauthorizedUsers!);
-      } catch (err) {
-        return res.status(500).send(err);
-      }
+        try {
+            const unauthorizedUsers = await userModel.find({isAuthorized: false});
+            return res.status(200).json(unauthorizedUsers!);
+        } catch (err) {
+            return res.status(500).send(err);
+        }
     });
-  return router;
+    return router;
 }
