+++ src/API/REST/routes/resend_verification_link.ts
@@ -1,7 +1,7 @@
 import express, { Router } from "express";
 import { tokenModel } from "../../../database/models/token";
 import {userModel} from "../../../database/models/user";
-import { send_verification_mail } from "../../../util/mailer";
+import {mailer} from "../../../util/mailer";
 
 export default function resend_verification_link(): Router {
     let router = express.Router();
@@ -13,22 +13,22 @@ export default function resend_verification_link(): Router {
             if (!email)
                 return res.status(400).send("Missing e-mail parameter.");
 
+            // check if user needs verification
             const user = await userModel.findOne({email});
-
             if (!user)
                 return res.status(404).send("Could not find user with this e-mail. Please register.");
-
             if (user.isEmailVerified)
                 return res.status(200).send("User has already been verified.");
 
+            // delete old token
             let token = await tokenModel.findOne({_userId: user._id});
-
             if (token)
                 await token.delete();
 
+            // create new token
             token = await tokenModel.create({ _userId: user._id });
 
-            send_verification_mail(email, token.token);
+            await mailer.send_verification_mail(user.firstName, email, token.token);
             res.status(200).send("Verification link has been resent.");
         }));
