+++ src/API/REST/routes/verify_email.ts
@@ -1,35 +1,32 @@
-import express from "express";
-import { userModel } from "../../../database/models/user";
-import { ExtRequest } from "../../../definitions/ext_request";
+import express, { Router } from "express";
+import { tokenModel } from "../../../database/models/token";
+import {userModel} from "../../../database/models/user";
 
-export default function verify_email_route() {
-  let router = express.Router();
+export default function verify_email(): Router {
+    let router = express.Router();
 
-  router
-    .get("/verify_email/:id", async (req: ExtRequest, res: any) => {
-      const userId = req.params.id;
-      const verificationToken = req.query.token;
+    router
+        .get("/verify/:token", (async (req, res) => {
+            const tokenObj =  await tokenModel.findOne({ token: req.params.token });
+            if (!tokenObj)
+                return res.status(404).send("Verification token could not be found. It may have expired.");
+                
+            const user = await userModel.findById(tokenObj._userId);
+            if (!user)
+                return res.status(404).send("Invalid token: User for this token could not be found.");
+            if (user._id)
+                return res.status(404).send("User for this token could not be found.");
 
-      if (verificationToken == null) {
-        return res.status(400).send("Missing token query parameter");
-      }
+            if (user.isEmailVerified) {
+                res.status(200).send("User has already been verified.");
+                return tokenObj.delete();
+            }
 
-      const user = await userModel.findById(userId);
+            user.isEmailVerified = true;
+            await user.save();
+            res.status(200).send("User Email has been verified");
+            tokenObj.delete();
+        }))
 
-      if (!user) {
-        return res.status(404).send(`User ${userId} not found`);
-      }
-
-      if (user.emailVerificationToken !== verificationToken) {
-        return res.status(400).send("Wrong token");
-      }
-
-      user.isVerified = true;
-      await user.save();
-
-      return res
-        .set("Content-Type", "text/html")
-        .send(Buffer.from("<h1>Email verified</h2><p>Congratulations, you have successfully verified your email address!</p>"));
-    });
-  return router;
+    return router;
 }
